+++ src/mv/gl/shape/sphere.cpp
@@ -12,8 +12,8 @@ namespace mv::gl::shape
         constexpr float u_start = 0.0F;
         constexpr float v_start = 0.0F;
 
-        constexpr float u_end = M_PI * 2.0F;
-        constexpr float v_end = M_PI;
+        constexpr float u_end = static_cast<float>(M_PI) * 2.0F;
+        constexpr float v_end = static_cast<float>(M_PI);
 
         const float u_step = (u_end - u_start) / static_cast<float>(slices);
         const float v_step = (v_end - v_start) / static_cast<float>(stacks);
